Create Day_4.py
Objective 
In this challenge, we're going to learn about the difference between a class and an instance; because this is an Object Oriented concept, it's only enabled in certain languages. Check out the Tutorial tab for learning materials and an instructional video!

Task 
Write a Person class with an instance variable, age , and a constructor that takes an integer, initialAge, as a parameter. The constructor must assign initialAge to  after confirming the argument passed as initialAge  is not negative; if a negative argument is passed as ,initialAge the constructor should set  Age to 0  and print Age is not valid, setting age to 0.. In addition, you must write the following instance methods:

yearPasses() should increase the age instance variable by .
amIOld() should perform the following conditional actions:
If , age<13 print You are young..
If  age>=13 and age<18 , print You are a teenager..
Otherwise, print You are old..